requirements.txt
absl-py==1.3.0
anyio==3.6.2
asgiref==3.5.2
astunparse==1.6.3
azure-core==1.26.0
azure-storage-blob==12.14.1
build==0.8.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
Django==4.1.1
django-environ==0.9.0
djangorestframework==3.14.0
flatbuffers==22.10.26
fonttools==4.37.3
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.7.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
isodate==0.6.1
joblib==1.2.0
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
msrest==0.7.1
multitasking==0.0.11
mysqlclient==2.1.1
numpy==1.23.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
pep517==0.13.0
Pillow==9.2.0
pip-tools==6.9.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
sniffio==1.3.0
sqlparse==0.4.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.2
urllib3==1.26.12
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
yfinance==0.1.74
